+++ notifications/tests/sample_notifications/apps.py
@@ -1,6 +1,0 @@
-from notifications.apps import Config as NotificationConfig
-
-
-class SampleNotificationsConfig(NotificationConfig):
-    name = "notifications.tests.sample_notifications"
-    label = "sample_notifications"