+++ sample_website/sample_app/apps.py
@@ -1,0 +1,8 @@
+""" Django notifications apps file """
+# -*- coding: utf-8 -*-
+from django.apps import AppConfig
+
+
+class Config(AppConfig):
+    name = "sample_app"
+    default_auto_field = "django.db.models.AutoField"