+++ sample_website/static/live-test.js
@@ -1,5 +1,3 @@
-console.log('running tester')
-
 function make_notification() {
     var r = new XMLHttpRequest();
     r.open("GET", '/test_make/', true);