+++ sample_website/urls.py
@@ -1,15 +1,14 @@
 """ Django notification urls for tests """
 # -*- coding: utf-8 -*-
 from django.contrib import admin
-from django.contrib.auth.views import LoginView
 from django.urls import include, path
 
-from notifications.tests.views import live_tester, make_notification
+from .views import live_tester, make_notification
 
 urlpatterns = [
+    path("__debug__/", include("debug_toolbar.urls")),
     path("test_make/", make_notification),
-    path("test/", live_tester),
-    path("login/", LoginView.as_view(), name="login"),  # reverse for django login is not working
     path("admin/", admin.site.urls),
     path("", include("notifications.urls", namespace="notifications")),
+    path("", live_tester),
 ]